More samples in test.cpp

+++ reduced_rational.cpp
@@ -1,53 +1,53 @@
 
 #include "reduced_rational.hpp"
 
-ReducedRational::ReducedRational() :
-Rational() {
-
+ReducedRational::ReducedRational() : Rational()
+{
 }
 
-ReducedRational::ReducedRational(const long& n) :
-Rational(n) {
+ReducedRational::ReducedRational(const double &n) : Rational(n)
+{
 	reduce();
 }
 
-ReducedRational::ReducedRational(const long& n, const long& d) : 
-Rational(n, d) {
+ReducedRational::ReducedRational(const long &n, const long &d) : Rational(n, d)
+{
 	reduce();
 }
 
-ReducedRational::ReducedRational(const char* cp) :
-ReducedRational(std::string(cp)) {
-
+ReducedRational::ReducedRational(const char *cp) : ReducedRational(std::string(cp))
+{
 }
 
-ReducedRational::ReducedRational(const std::string& str) :
-Rational(str) {
+ReducedRational::ReducedRational(const std::string &str) : Rational(str)
+{
 	reduce();
 }
 
-ReducedRational::ReducedRational(const Rational& other) {
+ReducedRational::ReducedRational(const Rational &other)
+{
 	this->numerator = other.numerator;
 	this->denominator = other.denominator;
 	reduce();
 }
 
-ReducedRational::ReducedRational(const ReducedRational& other) {
+ReducedRational::ReducedRational(const ReducedRational &other)
+{
 	this->numerator = other.numerator;
 	this->denominator = other.denominator;
 }
 
-ReducedRational::~ReducedRational() {
-
+ReducedRational::~ReducedRational()
+{
 }
 
-
-
-bool ReducedRational::isReduced() const {
+bool ReducedRational::isReduced() const
+{
 	return true;
 }
 
-void ReducedRational::reduce() {
+void ReducedRational::reduce()
+{
 	long i = gcd();
 	numerator /= i;
 	denominator /= i;
