+++ reduced_rational.hpp
@@ -3,22 +3,22 @@
 
 #include "rational.hpp"
 
-class ReducedRational : public Rational {
-private:
+class ReducedRational : public Rational
+{
+  private:
 	void reduce();
 
-public:
+  public:
 	ReducedRational();
-	ReducedRational(const long& n);
-	ReducedRational(const long& n, const long& d);
-	ReducedRational(const char* str);
-	ReducedRational(const std::string& str);
-	ReducedRational(const Rational& other);
-	ReducedRational(const ReducedRational& other);
+	ReducedRational(const double &n);
+	ReducedRational(const long &n, const long &d);
+	ReducedRational(const char *str);
+	ReducedRational(const std::string &str);
+	ReducedRational(const Rational &other);
+	ReducedRational(const ReducedRational &other);
 	~ReducedRational();
 
 	bool isReduced() const;
-
 };
 
 #endif
